pyproject.toml
[project]
name = "quake-legacy"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "httpx>=0.27",
  "pillow>=10.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "pyright>=1.1.360"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["creative_suite*"]
exclude = ["creative_suite.tests*", "creative_suite.engine.tests*"]

[tool.pytest.ini_options]
testpaths = ["creative_suite/tests"]
asyncio_mode = "auto"

[tool.pyright]
include = ["creative_suite"]
pythonVersion = "3.11"
typeCheckingMode = "strict"
